# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ExtractionResult": {
        "description": "Outcome of one feature-extraction run, embedded in the manifest.",
        "properties": {
          "device": {
            "title": "Device",
            "type": "string"
          },
          "embeddings_key": {
            "title": "Embeddings Key",
            "type": "string"
          },
          "encoder": {
            "title": "Encoder",
            "type": "string"
          },
          "feature_dim": {
            "title": "Feature Dim",
            "type": "integer"
          },
          "num_patches": {
            "title": "Num Patches",
            "type": "integer"
          },
          "patch_level": {
            "title": "Patch Level",
            "type": "integer"
          },
          "patch_size": {
            "title": "Patch Size",
            "type": "integer"
          },
          "tissue_fraction": {
            "title": "Tissue Fraction",
            "type": "number"
          }
        },
        "required": [
          "encoder",
          "device",
          "num_patches",
          "feature_dim",
          "patch_level",
          "patch_size",
          "tissue_fraction",
          "embeddings_key"
        ],
        "title": "ExtractionResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PresignedUpload": {
        "description": "A short-lived presigned PUT for a direct browser-to-B2 slide upload.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "headers": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Headers",
            "type": "object"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "method": {
            "title": "Method",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "key",
          "url",
          "method",
          "content_type",
          "headers",
          "expires_in"
        ],
        "title": "PresignedUpload",
        "type": "object"
      },
      "Slide": {
        "description": "Full manifest \u2014 the source of truth persisted as `manifest.json`.",
        "properties": {
          "bag_label": {
            "title": "Bag Label",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embeddings_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embeddings Key"
          },
          "encoder": {
            "title": "Encoder",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "feature_dim": {
            "title": "Feature Dim",
            "type": "integer"
          },
          "height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "level_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level Count"
          },
          "manifest_key": {
            "default": "",
            "title": "Manifest Key",
            "type": "string"
          },
          "mpp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mpp"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "num_patches": {
            "title": "Num Patches",
            "type": "integer"
          },
          "patch_grid_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Grid Key"
          },
          "patch_level": {
            "title": "Patch Level",
            "type": "integer"
          },
          "patch_size": {
            "title": "Patch Size",
            "type": "integer"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "source": {
            "enum": [
              "sample",
              "upload"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "source_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Key"
          },
          "status": {
            "enum": [
              "pending_upload",
              "registered",
              "extracting",
              "extracted",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "thumbnail_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Key"
          },
          "tissue_overlay_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tissue Overlay Key"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "width": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Width"
          }
        },
        "required": [
          "id",
          "label",
          "source",
          "status",
          "bag_label",
          "encoder",
          "patch_level",
          "patch_size",
          "created_at",
          "updated_at",
          "source_filename",
          "size_bytes",
          "size_human",
          "num_patches",
          "feature_dim"
        ],
        "title": "Slide",
        "type": "object"
      },
      "SlideCreate": {
        "description": "Form fields for registering a slide.\n\nFor `source=\"sample\"` no file fields are needed. For `source=\"upload\"` the\ndeclared `filename`/`content_type`/`size_bytes` let the API mint a presigned\nPUT so the browser streams the (multi-GB) slide straight to B2.",
        "properties": {
          "bag_label": {
            "default": "unknown",
            "maxLength": 60,
            "title": "Bag Label",
            "type": "string"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "encoder": {
            "default": "resnet50-truncated",
            "title": "Encoder",
            "type": "string"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "label": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "notes": {
            "default": "",
            "maxLength": 2000,
            "title": "Notes",
            "type": "string"
          },
          "patch_level": {
            "default": 0,
            "title": "Patch Level",
            "type": "integer"
          },
          "patch_size": {
            "default": 256,
            "title": "Patch Size",
            "type": "integer"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "source": {
            "default": "sample",
            "enum": [
              "sample",
              "upload"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "SlideCreate",
        "type": "object"
      },
      "SlideCreated": {
        "description": "Response to POST /slides.\n\n`upload` is populated only for the own-WSI upload path: the browser PUTs the\nbytes to it, then calls POST /slides/<id>/register. It is never persisted.",
        "properties": {
          "slide": {
            "$ref": "#/components/schemas/Slide"
          },
          "upload": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PresignedUpload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "slide"
        ],
        "title": "SlideCreated",
        "type": "object"
      },
...
      "SlideSummary": {
        "description": "Compact record for the Library grid (one per manifest.json).",
        "properties": {
          "bag_label": {
            "title": "Bag Label",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "encoder": {
            "title": "Encoder",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "feature_dim": {
            "title": "Feature Dim",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "num_patches": {
            "title": "Num Patches",
            "type": "integer"
          },
          "patch_level": {
            "title": "Patch Level",
            "type": "integer"
          },
          "patch_size": {
            "title": "Patch Size",
            "type": "integer"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "source": {
            "enum": [
              "sample",
              "upload"
            ],
            "title": "Source",
            "type": "string"
          },
          "source_filename": {
            "title": "Source Filename",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending_upload",
              "registered",
              "extracting",
              "extracted",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "thumbnail_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumbnail Key"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label",
          "source",
          "status",
          "bag_label",
          "encoder",
          "patch_level",
          "patch_size",
          "created_at",
          "updated_at",
          "source_filename",
          "size_bytes",
          "size_human",
          "num_patches",
          "feature_dim"
        ],
        "title": "SlideSummary",
        "type": "object"
      },
      "SlideUpdate": {
        "description": "Editable metadata. All optional; only provided fields change.",
        "properties": {
          "bag_label": {
            "anyOf": [
              {
                "maxLength": 60,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bag Label"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "title": "SlideUpdate",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for CLAM WSI Feature Extraction: a computational-pathology pipeline that tiles, tissue-segments, and CNN-embeds whole-slide images (WSI) with raw slides, patches, and feature tensors stored on Backblaze B2 over the S3-compatible API. This contract documents the local API, not a hosted public endpoint.",
    "title": "CLAM WSI Feature Extraction API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/slides": {
      "get": {
        "operationId": "list_slides_endpoint_slides_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SlideSummary"
                  },
                  "title": "Response List Slides Endpoint Slides Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Slides Endpoint",
        "tags": [
          "slides"
        ]
      },
      "post": {
        "operationId": "create_slide_endpoint_slides_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlideCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlideCreated"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Slide Endpoint",
        "tags": [
          "slides"
        ]
      }
    },
...
    "/slides/{slide_id}": {
      "delete": {
        "operationId": "delete_slide_endpoint_slides__slide_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Slide Endpoint",
        "tags": [
          "slides"
        ]
      },
      "get": {
        "operationId": "get_slide_endpoint_slides__slide_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Slide Endpoint",
        "tags": [
          "slides"
        ]
      },
      "patch": {
        "operationId": "update_slide_endpoint_slides__slide_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SlideUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Slide Endpoint",
        "tags": [
          "slides"
        ]
      }
    },
    "/slides/{slide_id}/asset/{name}": {
      "get": {
        "operationId": "slide_asset_endpoint_slides__slide_id__asset__name__get",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Slide Asset Endpoint",
        "tags": [
          "slides"
        ]
      }
    },
    "/slides/{slide_id}/extract": {
      "post": {
        "operationId": "extract_slide_endpoint_slides__slide_id__extract_post",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Extract Slide Endpoint",
        "tags": [
          "slides"
        ]
      }
    },
    "/slides/{slide_id}/register": {
      "post": {
        "operationId": "register_slide_endpoint_slides__slide_id__register_post",
        "parameters": [
          {
            "in": "path",
            "name": "slide_id",
            "required": true,
            "schema": {
              "title": "Slide Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Slide"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register Slide Endpoint",
        "tags": [
          "slides"
        ]
      }
    },
...
  }
}